+++ studio/src/app/pages/editor/app-editor/app-editor.tsx
@@ -9,7 +9,7 @@ import busyStore from '../../../stores/busy.store';
 import authStore from '../../../stores/auth.store';
 import colorStore from '../../../stores/color.store';
 
-import {debounce, isFullscreen, isIOS, isMobile} from '@deckdeckgo/utils';
+import {debounce, isAndroidTablet, isFullscreen, isIOS, isIPad, isMobile} from '@deckdeckgo/utils';
 
 import {convertStyle} from '@deckdeckgo/deck-utils';
 
@@ -596,7 +596,7 @@ export class AppEditor {
   }
 
   private initMainSize() {
-    if (!this.contentRef || isFullscreen() || isMobile()) {
+    if (!this.contentRef || isFullscreen() || (isMobile() && !isIPad() && !isAndroidTablet())) {
       this.mainSize = {
         width: isMobile() ? 'calc(100% - 32px)' : '100%',
         height: isMobile() ? 'calc(100% - 32px)' : '100%',
@@ -606,7 +606,9 @@ export class AppEditor {
 
     const maxHeight: number = this.contentRef.offsetHeight - 32;
 
-    const width: number = this.contentRef.offsetWidth - 32;
+    const wideScreen: MediaQueryList = window.matchMedia('(min-width: 1200px)');
+
+    const width: number = this.contentRef.offsetWidth - (wideScreen.matches ? 192 : 32);
     const height: number = (width * 9) / 16;
 
     this.mainSize =
@@ -717,54 +719,65 @@ export class AppEditor {
     this.activeIndex = index;
   }
 
+  private async selectStep($event: CustomEvent<HTMLElement>) {
+    this.actionsEditorRef?.selectStep($event?.detail);
+  }
+
   render() {
     const autoSlide: boolean = deckStore.state.deck?.data?.attributes?.autoSlide !== undefined ? deckStore.state.deck.data.attributes.autoSlide : false;
 
     return [
       <app-navigation publish={true} class={this.hideNavigation ? 'hidden' : undefined}></app-navigation>,
-      <ion-content ref={(el) => (this.contentRef = el as HTMLElement)}>
-        <main
-          ref={(el) => (this.mainRef = el as HTMLElement)}
-          class={busyStore.state.slideReady ? (this.presenting ? 'ready idle' : 'ready') : undefined}
-          style={{'--main-size-width': this.mainSize?.width, '--main-size-height': this.mainSize?.height}}>
-          {this.renderLoading()}
-          <deckgo-deck
-            ref={(el) => (this.deckRef = el as HTMLDeckgoDeckElement)}
-            embedded={true}
-            style={this.style}
-            reveal={this.fullscreen && this.presenting}
-            direction={this.direction}
-            directionMobile={this.directionMobile}
-            animation={this.animation}
-            autoSlide={this.fullscreen && this.presenting && autoSlide ? 'true' : 'false'}
-            onMouseDown={(e: MouseEvent) => this.deckTouched(e)}
-            onTouchStart={(e: TouchEvent) => this.deckTouched(e)}
-            onSlideNextDidChange={() => this.onSlideChange()}
-            onSlidePrevDidChange={() => this.onSlideChange()}
-            onSlideToChange={() => this.onSlideChange()}>
-            {this.slides}
-            {this.background}
-            {this.header}
-            {this.footer}
-          </deckgo-deck>
-          <deckgo-remote autoConnect={false}></deckgo-remote>
-          <app-slide-warning></app-slide-warning>
-        </main>
+      <ion-content class="ion-no-padding">
+        <div class="grid">
+          <div class="deck" ref={(el) => (this.contentRef = el as HTMLElement)}>
+            <main
+              ref={(el) => (this.mainRef = el as HTMLElement)}
+              class={busyStore.state.slideReady ? (this.presenting ? 'ready idle' : 'ready') : undefined}
+              style={{'--main-size-width': this.mainSize?.width, '--main-size-height': this.mainSize?.height}}>
+              {this.renderLoading()}
+              <deckgo-deck
+                ref={(el) => (this.deckRef = el as HTMLDeckgoDeckElement)}
+                embedded={true}
+                style={this.style}
+                reveal={this.fullscreen && this.presenting}
+                direction={this.direction}
+                directionMobile={this.directionMobile}
+                animation={this.animation}
+                autoSlide={this.fullscreen && this.presenting && autoSlide ? 'true' : 'false'}
+                onMouseDown={(e: MouseEvent) => this.deckTouched(e)}
+                onTouchStart={(e: TouchEvent) => this.deckTouched(e)}
+                onSlideNextDidChange={() => this.onSlideChange()}
+                onSlidePrevDidChange={() => this.onSlideChange()}
+                onSlideToChange={() => this.onSlideChange()}>
+                {this.slides}
+                {this.background}
+                {this.header}
+                {this.footer}
+              </deckgo-deck>
+              <deckgo-remote autoConnect={false}></deckgo-remote>
+              <app-slide-warning></app-slide-warning>
+            </main>
+          </div>
+
+          <app-breadcrumbs slideNumber={this.activeIndex} onStepTo={($event: CustomEvent<HTMLElement>) => this.selectStep($event)}></app-breadcrumbs>
+
+          <app-actions-editor
+            ref={(el) => (this.actionsEditorRef = el as HTMLAppActionsEditorElement)}
+            hideActions={this.hideActions}
+            fullscreen={this.fullscreen}
+            slides={this.slides}
+            slideNumber={this.activeIndex}
+            onSignIn={() => this.signIn()}
+            onAddSlide={($event: CustomEvent<JSX.IntrinsicElements>) => this.addSlide($event)}
+            onAnimatePrevNextSlide={($event: CustomEvent<boolean>) => this.animatePrevNextSlide($event)}
+            onSlideTo={($event: CustomEvent<number>) => this.slideTo($event)}
+            onSlideCopy={($event: CustomEvent<HTMLElement>) => this.copySlide($event)}
+            onElementFocus={($event: CustomEvent<HTMLElement>) => this.onElementFocus($event)}
+            onPresenting={($event: CustomEvent<boolean>) => this.updatePresenting($event?.detail)}></app-actions-editor>
+        </div>
         <app-slide-preview deckRef={this.deckRef}></app-slide-preview>
       </ion-content>,
-      <app-actions-editor
-        ref={(el) => (this.actionsEditorRef = el as HTMLAppActionsEditorElement)}
-        hideActions={this.hideActions}
-        fullscreen={this.fullscreen}
-        slides={this.slides}
-        slideNumber={this.activeIndex}
-        onSignIn={() => this.signIn()}
-        onAddSlide={($event: CustomEvent<JSX.IntrinsicElements>) => this.addSlide($event)}
-        onAnimatePrevNextSlide={($event: CustomEvent<boolean>) => this.animatePrevNextSlide($event)}
-        onSlideTo={($event: CustomEvent<number>) => this.slideTo($event)}
-        onSlideCopy={($event: CustomEvent<HTMLElement>) => this.copySlide($event)}
-        onElementFocus={($event: CustomEvent<HTMLElement>) => this.onElementFocus($event)}
-        onPresenting={($event: CustomEvent<boolean>) => this.updatePresenting($event?.detail)}></app-actions-editor>,
       this.renderInlineEditor(),
     ];
   }