+++ utils/utils/CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="1.8.0"></a>
+
+# 1.8.0 (2021-01-30)
+
+### Features
+
+- `isAndroid`
+- `isAndroidTablet`
+- use `fullscreenElement` (with prefix) to detect full screen
+
 <a name="1.7.0"></a>
 
 # 1.7.0 (2020-11-21)